route prices to pdf & new button bug fix

+++ Heka.DataAccess/Context/Models/ItemOfferDetail.cs
@@ -37,7 +37,7 @@ namespace Heka.DataAccess.Context
         public Nullable<decimal> ProfitRate { get; set; }
         public Nullable<int> CreditMonths { get; set; }
         public Nullable<int> CreditRate { get; set; }
-        public Nullable<int> SheetTickness { get; set; }
+        public Nullable<decimal> SheetTickness { get; set; }
 
         [ForeignKey("ItemOfferSheet")]
         public Nullable<int> ItemOfferSheetId { get; set; }
